pyproject.toml
[project]
name = "chatterbox-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chatterbox-tts>=0.1.6",
    "onnx<1.17.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
]
